Add globbing for gem files
The gem file list has to be an explict list of every
file to be added. This commit adds a method to allow
us to keep the existing project_data.yaml format of
a file list including just directory names by globbing
in the task itself. This also excludes the packaging
repo from the gem file list.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>