Add find_tool, use sh instead of %x in shell calls
This commit adds the find_tool, which checks for the existence of a tool
and returns its location. This allows the tasks to prefer tools and fall
through without failing on the first non-existent tool as would happen
with check_tool. It also moves a few calls from using %x to using sh, as
%x supresses STDOUT, which is helpful when a password prompt is given so
the user knows why the script has paused.