Fixup versioning for release and format
We started using dashes after the rc in our tags,
e.g. puppet-2.7.19-rc1 but the version determination
in packaging hasn't caught up. This adds support
for the dash version. The old method is kept because
it is required by rubygems. This commit also resolves
an issue where '-1puppetlabs1' is not added to the
deb version string if building an rc. Finally this
commit splits get_release out into rpm_release and
deb_release to avoid using an rpm_release string
in a deb.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>