fix debian version for non-tag git describes
The git describe version string that does not contain
an rc in it breaks debian tarball building because it
expects an orig tarball to match the odd different
version string parsing by dashes. This commit ensures
that the github describe is joined by dots in the
deb version to compose a correct version string.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>