change target version to 10.4 and fixup build time
This commit changes the target version when building apple
packages to 10.4, which changes the package format from a
"flat" package back to the older package format. This
addresses issues seen on os x 10.5.8 where packages can
behave in unexpected and undesirable ways. This format
is still compatible with current (lion) systems, and can
probably be kept until we formally deprecate os x 10.5.x
support. The date format is modified for compatibility with
the older package info plist format.