(packaging) add signing and shipping tasks for solaris
This patch adds signing and shipping tasks to solaris ips. We need atleast
a self signed certificate and its key to use the task in pem format.

This patch also assumes that the user that built the package in the current
system is present in the ips repository and we have permissions to sync.