Add a re-try option, initially to cow-building
currently, building for a set of cows is one of the
longest packaging tasks, and a failure in any one
essentially requires the whole set be redone. This
commit adds a rescue to the cow building method that
pauses the rake task and prompts the user for a 'retry'
if desired, allowing us to fix any issues with the
current cow and pick up where we left off. Since its
a utility function, this can be added to other areas
in packaging as well.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>