Update cows before building
Previously the cow would be in an unknown state before building. This commit
chains the update_cow method to the cowbuilding task. It also updates the
user's path to include /usr/sbin, as that may not be present but is needed to
find cowbuilder and its associated tools. The call to cowbuilder is also fully
qualified, and is called via `sudo -E` to preserve the current environment in
the sudo call.