Update mock task to correctly link noarch rpms
Previously the mock method would switch on the full path of the rpm, which
would never include noarch. This commit addresses that by switching on
File.basename(rpm) which will limit the switch to the filename, which will have
noarch, i386, x86_64, or src in its contents. It correctly links noarch
packages into i386 and x86_64.