cleanup boolean conditionals with a helper method
This commit introduces a helper method, boolean_value
to sanitize input of boolean values from the yaml files
and ENV variables. Various tasks are updated to use it.
This commit also introduces @build_ips, a value we
can use to specify if a product supports building solaris
packages or not.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>