Add remote ship of code repo with packaging bootstrap
This commit adds a method that generates a bundle in
a temporary location with tags from the current repository
at a specified commit. It also adds a method that ships
this bundle to a temporary location to a remote host,
unpacks/clones it, and bootstraps the packaging repo.
A helper method that generates a small random string
is added for naming the bundles, to reduce risk of
collision on the remote builder host.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>