Add task to fetch additional build data
This commit adds an additional task that fetches
additional data from a builder_data.yaml file that
augments automation of packaging but is not really
pertinent to the mainline packaging repo. It loads
more variables, and the presence of the builder_data
file is used to conditionally present additional tasks
specific to puppet labs infrastructure.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>