Add gem name, description and summary support
In some cases, the gem name, description and/or summary may be
different for a gem package in a repo than the other packaging
forms, e.g., if only a client-side is built as a gem. This commit
adds support for these settings as well in the project_data yaml
file.