Update cow in the right place
Currently, the update_cow call happens in the :build_deb
task, which breaks if using package:deb. This is because
:build_deb is a super-task that calls to the different
building methods, e.g cows vs not cows, depending on what
arguments its given, and we don't have everything set to
run a cowbuilder update in the dpkg-buildpackage call.
This commit moves the update call into the pdebuild method,
where its always and only applicable.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>