make the build_rpm_with_mock method saner
By adding a mocks parameter to build_rpm_with_mock, we make
it a bit more isolated and intuitive. This commit then
leverages this to add a pl: default_mock task, e.g. and fixup
the pe: default mock task to use it instead, as exists already
on the debian side. A @default_mock variable is added that just
defaults to the first mock in the mocks list.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>