Add auto-remote repo update on PE ship and fix path
This removes the hardcoded repo path in pe shipping,
and adds repo updating as a post-task of shipping.
It changes shipping so that pe-dev packages are always
shipped as unified assuming the default cow as the build.
Finally, this commit update pe shipping to use pe-the-
things instead of deb-the-things.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>