Use monotonically increasing versions for pe dev building
This commit changs the version string when building a pe dev project to using
the commit number as part of version, e.g. what is suppled by the utility
method #get_dot_version. For example, given a git describe of 1.1.7-23-<hash>
this commit forms the version 1.1.7.23. This allows the version to continue to
increase as the commits increase, but it does present the difficulty that final
versions will be 'lower' than the previous build versions, but this can be
solved by tagging 'up', e.g. for builds at 1.1.7.23 just tag as 1.1.8, or by
keeping separate package repos/workflows entirely for enterprise production
and pe devs.