Fix the macro for %dist detection on rpm platforms.
Prior to this commit, when building an RPM on EL6, the dist tag would
still be tagged with el5.  This commits dynamically detects the dist
tag.

Dist is provided on all EL and Fedora systems except EL5.  The solution
is to query if it's EL and then see what the major version of the RPM is
that provides that file.

Fixes #50 Pull request as well.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>