Split build extras data out from fetch
This commit breaks the pl:fetch task into two tasks, one
that retrieves the extras file and the other that loads the
variables. By doing so, we can use the load extras task to
override settings in build_defaults without having to fetch
the file every time. This commit also adds support for overriding
the yum and apt repo paths, and for an extra freight conf setting.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>