Add sles building for PE to the packaging repo
This commit adds support for building sles packages for PE
using the pe repo. Several variables are added to extras data
to support this. Currently this is PE only, but once some more
planning is done around sles we can expand to foss as well.
This commit includes shipping of sles as well. This is very
hacky and terrible, but works for now.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>