Use standard Rake FileList exclusion patterns in tar
The tar package target would build the tarball using, typically, a recursive
copy of several top level directories in the project.  That actually ends up
being a pain in the neck, as it also copied editor backup files and other junk
into place.

This replaces that with a Rake FileList, and iteration.  That means we perform
more individual copy operations, but we get the benefit of the existing and
well tested code to exclude undesirable things from the target directory.

Listing directories does the right thing, however, implicitly expanding into a
pattern of `directory/**/*` to install all contained files.  That makes many
project installation lists much, much cleaner.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>