Copy files to workdir quietly instead of using tar
Previously the tar task passed the list of files to be installed in a tarball
on the commandline to tar and piped it to tar x to get a list of files. That
would fail on large lists of files. This commit addresses that by looping over
all of the files and copying each file into the correct path in the workdir.
Verbose is set to false for both calls, so the output should be minimal.