Handle empty directories in tar task
Currently the package:tar task will not include empty directories in the
tarball. This commit adds support for these. An 'empty_dir?' utility method is
added which just checks if a file exists, is a directory, and is empty. If a
directory in the list is empty, one of the same name is created in the tar
structure, to ensure hidden sub-files such as .gitignore aren't copied over
with "empty" directories as would happen with cp_r. If not an empty directory,
the existing logic is essentially followed.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>