Only create the 'pkg/rpm' directory if using rpmbuild
Currently the package:srpm task (a prereq to mock builds) creates a 'pkg/rpm'
directory even though no rpm is getting built, only a srpm. When a mock rpm
build is run, the same is true, even though the final rpms will be placed in
a different structure. This commit updates the build_rpm method to only make
a 'pkg/rpm' directory if building an rpm using the package:rpm task.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>