Add test for -rcX versions, fix git/ips version for this as well
Per semver, we separate the rcX string from the major.minor.release version
with a dash, e.g. 2.7.20-rc1. This commit augments the version spec test to
ensure the appropriate versions are returned with a git tag of this format.
In updating the tests, it was discovered that the git_describe_version and
get_ips_version methods don't handle this case quite right, because in
the case of a -rcX, the describe string will contain 4 significant
elements, rather than 3. Here we fix this up as well.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>