Add metrics.rake file for metrics keeping
This commit is the first of several steps towards adding metrics keeping on
package builds for Puppet Labs to the packaging repo. It adds a Metrics_helper
module that server as the entry point for posting metrics on a remote postgres
server. Much of the connection info is assumed to come from environment vars.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>