Add basic uploading of ips packges to downloads
This commit adds a couple optional variables to the build-extras file
for uploading ips p5p packages to the solaris downloads. It also adds
a task for uploading built p5p packages there. This allows us to make
these packages available for those who might use them until we can
evaluate other means of distribution. It also fixes the rsync call
to the existing repo update call for ips so that we only ship files
rather than the pkgs directory.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>