Add new version update task with auto-commit
This commit adds another task to be called post-release that updates and
commits the version file with a required VERSION parameter. It commits
the change automatically. It also updates the existing version file
update task to use an optional env var for version, so we can call it from
the new one.