Add signing for IPS packages
This commit fixes IPS signing. It adds support for passing in the various cert
locations, and adds signing as a task. It renames the placeholder call to the
signing task to the correct name, 'pl:sign_ips' since we'll be settling on that
for the task name. It reworks the signing task, updating it to receive arguments
for repo and package fmri and sign the packagin in-repo, as opposed to the p5p
package itself, which wasn't correct.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>