Use %x rather than sh for pkglint
ips pkglint exits non-zero for dependency warnings, causing rake to exit. We
should use a %x call instead for this to get output but not necessarily exit
because of the warnings. This is sort of a compromise because it relies on
the user to be aware of errors that pkglint generates, but its needed for the
basic case of our packaging.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>