Rework namespaces
Because the ips variables reside in the package:ips namespace, we can't
reference them externally from the package:ips task (for passing along
signing arguments). This commit moves the bulk of the package:ips task
inside the package:ips namespace, and renames it :p5p. This more accurately
reflects its purpose anyhow. For backward compatibility, the package:ips
task is kept, and changed to reference the new :p5p task. The pl:ips task
is updated as well.