Add support for other tars in remote builds
Solaris, sadly, offers our required functionality only through gtar. This commit
updates the remote_bootstrap method to take an optional argument of the tar to
use on the remote host, and updates the remote_build task to accept/use this as
well. Then in the remote_ips task, we pass in 'gtar' as the tar to use.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>