Allow remote building without signing
This commit adds tasks that allow remote building of packages without signing.
This will be useful in automating the CI pipeline, allowing us to build from a
central location with jenkins without necessarily acknowledging release-able
status by signing and requiring a person on the other end. It also corrects an
error in one of the remote build task descriptions.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>