Add tests for expected debversion, rpmversion, rpmrelease
This commit adds tests for the expected values of debversion, rpmversion, and
rpmrelease. When making commits off of an rc tag, the # of commits past the tag
becomes part of the release. When making commits off of a non-rc tag, the # of
commits past the tag becomes part of the upstream version. These tests
currently fail, but the next commit fixes them.