Add another snowflake version of burning in package versions
Like every other project, Razor clearly needs a uniquely shaped method to burn
in version numbers during release.  This adds my own triangle to the square,
the rhombus, and the other methods already found here.

In practice, the others all assume that the "this is not versioned" string
uniquely occurs in the file, which makes them all unsuitable for my "extract
the version from git" code.  Alas.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>