Integrate version burn-in to the tar build process
In the longer term, having to commit a modified version file back to source
control means more windows for human error breaking something, as will as more
challenges around automating builds.

This adds a `NEW_STYLE_PACKAGE` option that integrates the version bump
process into the tar build - so that it only modifies the temporary working
copy of the code, not the master repository.

This allows Razor to have upstream tags be the sole source of version truth,
eliminating one of the typical copies of that data stored in git.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>