Unify rc and final tasks, deb and mock task naming
This commit introduces a deprecate utility method that just prints a
deprecation warning. It deprecates "rc" and "final" tasks in favor of tasks
that handle both now that the devel repos aren't necessary for dependency
resolution when building. The old tasks are hidden, but remain functional for
now. Remote tasks have been cleaned up so that their names match the local
tasks, just under a :remote namespace.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>