Fix/remove unnecessary subdir assignments
First, this commit changes the assignment of subdir inside the
build_rpm_with_mocks method to use the argument passed in rather than call out
to the is_rc? method. Second, it removes all the unnecessary assignments of
subdir inside the rake tasks, which used to be necessary but are no longer.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>