Split out build-data into team and project-specific files
Currently, the build-data files are a mess. This commit take the single
monolithic file for each team and splits it into two, one for team-specific
data, and another for project-specific data. This will be important in allowing
us to proceed with dual-build projects (e.g. FOSS and PE from the same source).

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>