Use same versioning across pe/foss
Before our versioning problems had been resolved, a quick workaround for PE
builds of branches with commits past a tag was to use the dot version instead
of a dash version. This isn't necessary any longer and is biting us now in
puppetdb. This commit just removes the conditional logic and uses the dash
version across foss/pe, which was the case originally.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>