Make project_data and team_data truly optional
One of the awesome things about ruby hashes, is that nil isn't a hash. So if
you try to grab a value from nil using a key, ruby complains. For example
@project_data['pe_name'] fails horribly if @project_data is nil and not {}. So
this commit makes those instance variables {} if they were nil, so that these
lookups don't cause exceptions, but just return false or nil themselves.