Ship debs to incoming/unified and rsync out of dist
Currently the pe:all task calls pe:deb followed by pe:ship_debs (and ships to
disparate). This means that when packages are build in /pkg/pe/deb/squeeze,
they get rsynced to repos/incoming/disparate and then they also get shipped to
only the squeeze repo. This commit changes where they ship to unified and
doesn't ship the $dist directory with the ship. This ensures that the package
gets freighted to all the dists for the PE version.