Remove test for root in building apple packages, use sudos
We should call 'sudo' to do everything we need sudo for, so checking that we're
running as root is not necessary, as it makes generic automation incredibly
difficult. This commit updates the script to do so. It also removes a strange
recursive chmod to 755 of the entire pkg directory. I'm fairly certain we don't
want to do this. Given the strange behavior we've experienced with apple
packaging, packages have been built and fdiffed before and after this commit to
ensure they are identical.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>