Randomize location of remote params file drop
As @haus pointed out, the destination for the remote params file is static, and
will lead to collision if more than one build of the same commit is invoked
targeted at the same builder. This commit places the file in a pseudo-random
subdirectory of /tmp rather than /tmp. Originally, this file was randomly named
so we could reasonably expect to avoid collisions, but when we changed to
git-sha, this was no longer the case.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>