Remove setup_el_dirs method
The setup_el_dirs task used hardcoded paths to make directories in advance of
mocking the packages. This was fragile and prone to breakage. This commit moves
the directory creation into the build_rpm_with_mock method where the current
destination is known and can be created before any copying happens.