Don't attempt to create rpm repo configs if there are none
We don't want to try to create rpm repo configs from a file that is
non-existent in the case that no rpms exist. This commit updates the task to
just exit if no rpm repos exist to create configs for. It also updates the task
to check for "repodata", which is a much better indicator of whether repos
exist than just "rpms".

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>