Link "pl:fetch" and "pl:load_extras" to remove static extras
Currently, the pl:fetch task creates $HOME/.packaging/{team,project} and
downloads two yaml files into these directories. This is dirty, both because
we're creating files on user's systems that they don't know about, and because
we're leaving them behind afterwards. This commit addresses this behavior by
downloading the files into a temporary directory instead. The static directory
is currently necessary because pl:fetch and pl:load_extras are independent and
don't pass data to each other. With the change to a random directory, this
commit makes pl:load_extras a hard dependency of pl:fetch, so that pl:fetch can
pass the directory containing the files to pl:load_extras. This commit removes
all invokations of pl:load_extras as a task in the packaging repo, replacing
them with pl:fetch.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>