Sleep uber tasks to avoid jenkins dropping jobs
When jenkins things we're DOSing it by posting to the same job 5 times in 2
seconds, it drops some of the jobs. In repeated testing, sleep 5 was the
shortest amount of time in which no jobs were dropped during the newly
parallelized mock_all and deb_all, so it stands to reason that the same would
apply to the uber tasks.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>