Always remove lock file even if repo creation fails
Currently if the repo creation fails in some ways, the lock file will not get
removed, stalling the repo creation until it is removed manually. This commit
updates this process to separate the lockfile removal into its own command,
which will be run regardless of the success or failure of the repo creation.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>