Update deb repos to use wget rather than sshing into distribution server
Currently, generating repo configs for debian requires sshing into the
distribution server and composing a list of available repos. This is not ideal
because not everyone will have the appropriate access. This commit updates the
deb_repo_configs task to use the internally "public" repos with wget instead,
so that anyone on the network can create the configs without special access.
While this was certainly hacky before, it is now still quite hacky. This uses
wget to generate a directory listing of the proposed hash's repos, and then
parse them into usable output with various array and string transformations.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>