sanify rpm repo config creation, don't try to ssh into distribution server
Currently, generating repo configs for red hat requires sshing into the
distribution server and composing a list of available repos. This is not ideal
because not everyone will have the appropriate access. This commit updates the
rpm_repo_configs task to use the internally "public" repos with wget instead,
so that anyone on the network can create the configs without special access.
While this was certainly hacky before, it is now still quite hacky. This uses
wget to generate a directory listing of the proposed hash's repos, and then
parse them into usable output with various array and string transformations.