Add support for pre_tar_tasks
This commit pulls in PR 118 from @jeffweiss
(https://github.com/puppetlabs/packaging/pull/118) which adds a tasks that
puppet dashboard will need to execute prior to composing its tarball. The
pre-tar task execution can actually be quite useful for other projects. So this
commit adds a simple conditional to check for the existence of the pre_tar_task
attribute in the build object (which will be added by the build_defaults or
project_data yaml files), and then executes the task prior to composing the
tarball.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>