Update cow-deb shipping to all cows
Currently we just assume the default cow was built and ship to 'unified', but
we swapped workflows awhile back to build all cows. This means we were just
missing most of the dists in the actual freight. Now, we just ship to disparate
with all the different distributions we built for.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>