Add pe prefix to deb repo creation
For pe, the deb packages are staged under the pe directory, and the deb repo
creation task was hardcoded to use just deb. This commit conditionally adds a
pe prefix to pe builds. This really just makes the hard coding a little better,
and could probably be abstracted in a better way.