Split repo file generation, shipping, and retrieval into separate tasks
Currently repo config file generation generates new repo configs at runtime on
the local station. This is not optimal because it means to get repo configs you
have to have $project cloned with the packaging repo to get repo configs. This
commit changes the workflow by creating the repo configs in one step, shipping
them to a static (and predictable) location on the distribution server, and
modifying the current rpm_repo_configs and deb_repo_configs to just retrieve
the stored configs on the server. This allows us to just wget the configs since
we know what their location will be. We can generate the configs and store them
on the distribution server as a post-repo generation step.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>