Add support for the build manager
This adds a number of `pe:jenkins:managed_*` and `pl:jenkins:managed_*
tasks to interact with the build manager. These jobs are wrappers
around the non-managed tasks of similar name which initialize the
build manager for a new build, and then call the underlying task with
some extra parameters set in order to ensure the various packaging
runs know they are managed.

It also modifies the `pl:deb`, `pl:mock`, `pl:gem`, `pl:dmg`,
`pl:tar`, and `pe:local_sles` tasks so that they will check in to the
build manager on completion if a managed build is running.