vendor gems without installing
Prior to this commit, the call to bundler would cause the gems to be
installed, thereby compiling native extensions for the as well, prior to
being vendored as plain ol' .gem files. This commit implements
functionality to look at the Gemfile and Gemfile.lock to determine
exactly which gem versions to install, excludes anything wholly
contained in the :development or :test groups, then then does an api
call to execute `gem fetch <gemname> -v <version>`.