Clean up package_repos before rsync call
Having multiple versions of the same package in the package_repos directory
causes problems when building. Sles' build automation is not smart about
choosing packages, so it is important to clean up the package directory before
rsyncing into it. This commit adds an rm_rf call to clean up all the files in
~/package_repos before building.