Add 00_monkey_patches.rake with Rake::Task#unshift method
This commit adds 00_monkey_patches.rake, which will be a file containing any
monkey patches we need. Initially, this contains Rake::Task#unshift, which is a
method that is based of the existing Rake::Task#enhance method. Rake task
objects contain an @actions variable, which is an array of blocks that are
executed when the #execute method is called. These blocks are normally created
by the common `task :name` syntax. Under the hood, Rake calls #enhance on the
task, which builds the task by appending a block passed as an argument to
tasks workflow, which is not to append to execution blocks, but rather prepend.
That is, we want a block to fire off before a tasks existing actions. This is
what the #unshift method provides. Similar to Array#unshift, this method
inserts a block at the beginning of the array of blocks to be executed by a
rake task.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>