Add support for modifying tasks on the fly for build manager
Modify existing Rake tasks with behavior to support managed builds Use
monkey-patched #unshift method and existing #enhance method to enable
inserting actions into already defined rake tasks on the fly.

For core tasks (e.g. pl:deb, package:tar, etc) we use #enhance to add an
additional execution block after the existing tasks, containing a status post
to the build server. Each individual task (e.g. deb cow build) has to post a
status, which is why we're adding at this level. Once the task finishes it posts to the build manager that it has completed.

For aggregate-style task, e.g. "pl:jenkins:uber_build" we prepend to the
existing execution blocks with one which notifies the build manager that a
build has started. Prepending to existing execution blocks offers the ability
to retain all existing task behavior including order in dependencies.

The downside of this approach over @branan's approach is that there is no
explicit failure state from a failed package task, because if a task throws an
exception in one block, Rake will not continue with the remaining blocks. This
means a failure will always mean a timeout, because the core task will never post a status if it fails.

The upside of this approach is that it is much cleaner (for varying definitions
of clean) in that does not add more tasks or clutter to the packaging repo, and
is largely transparent to all existing users. Just add MANAGED=true and the
rest is taken care of for you.